pyproject.toml
[tool.poetry]
name = "api-quick-start"
version = "0.1.0"
description = ""
authors = ["JB-Tellez <jb.tellez@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
djangorestframework-simplejwt = "^4.6.0"
whitenoise = "^5.2.0"
django-cors-headers = "^3.7.0"
gunicorn = "^20.0.4"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
black = "^20.8b1"
requests = "^2.25.1"
fire = "^0.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
